Changed ScelPeriodicSymCompare to accept matrix in constructor

+++ src/casm/clex/Supercell.cc
@@ -162,8 +162,8 @@ namespace CASM {
     return volume() * basis_size();
   }
 
-  Eigen::Matrix3i Supercell::transf_mat() const {
-    return this->sym_info().transformation_matrix().cast<int>();
+  Eigen::Matrix3l Supercell::transf_mat() const {
+    return this->sym_info().transformation_matrix();
     /* return iround(this->sym_info().transformation_matrix()); */
     //return CASM::transf_mat(primclex().prim().lattice(), lattice(), primclex().crystallography_tol());
   }
