+++ tests/unit/crystallography/BasicStructure_test.cpp
@@ -304,7 +304,7 @@ TEST(BasicStructureSiteTest, IsPrimitiveTest) {
 
   for(auto it = scel_enum.begin(); it != scel_enum.end(); ++it) {
     Eigen::Matrix3l transformation_matix = xtal::make_transformation_matrix(prim.lattice(), *it, prim.lattice().tol());
-    BasicStructure super = xtal::make_superstructure(prim.lattice(), transformation_matix.cast<int>());
+    BasicStructure super = xtal::make_superstructure(prim.lattice(), transformation_matix);
 
     EXPECT_EQ(super.lattice().is_right_handed(), true);
